add category page
This introduces CategoryPage as a new page type and will represent a page
with a list of articles. These articles can be from one category or all
articles of the system.
Currently it is only implemented for the view of all articles.

This also introduces a new theme file 'category' for this view.