Updating images and thumbnails
The main focus for this update is on images and the thumbnails created for it.
Before this, thumbnails were created in one size only and the naming convention
was quite weird.

Now new uploaded images keep their name, and newly created thumbnails will obtain
the same name with the addition of the size at the end of it.

This update also requires a database update, which can be done by running the
'update_images.php' script. (Reminder to myself: This should actually be in some
versioning/update process.)

For now, two thumbnail sizes will be created upon image upload (one for each theme
included in this repository), but new sizes can easily be created with the
getPathThumb() method by providing width and height to it.

The future plan for images is, to decouple them from the articles itself and
introduce a media library in the backend. This would make the use of images a
lot more flexible.

Also included: Bug fixes and minor refactoing.