Finish downloads update script
Instead of being something completely separate, download files are now considered
attachments to an article. They're also completly independent and don't belong to
only one specific article.

The existing download information pages are being transformed into actual articles
and connected to a 'Downloads' category. Unfortunately they will have the current
date assoziated with them and therefor appear at the beginning of your main page.
This is due to the lack of a date field for downloads.

In order to being able to access the files from the front end, themes are adjusted
so that they display an attachments section. This section is listing all attachments
with additional information about version, license and download counter of an
attachment.

The next commits will handle the removal of code that's no longer needed, as well
as a redesign of the admin pages for downloads. This commit also temporarily disables
the download counter, this will be fixed in an upcoming commit as well.