big admin backend refactor
I know this commit is probably way bigger than it should be, but it's a
complete rework of all admin pages.

It gets rid of a lot of legacy code, introduces I18n to pages were it
was missing and is more consistent with the page concept.

The pages to manage downloads are removed temporarily, since it wouldn't
make sense to transfer them as they are. Since downloads are simple
articles now with file attachments, the options to attach and manage
those files will be moved into the article edit/create pages.